Update patch set 1
Patch Set 1:

I don't really understand the issue.
When we call task_wake(), we set a bit in the events bitmap of the task, later we reach the task_wait_event() call, it will return immediatly as there is an event pending.
Not sure why we need to re-implement the event management with the has_event_pending variable.
Is something resetting the event bit during the task processing ? if so we can use another event bit instead of the default one.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0