Retry ectool commands that return EC_RES_IN_PROGRESS
This came up when testing issue 242706. If you tell the EC to return
EC_RES_IN_PROGRESS, then the EC gets stuck and times out on the next
command.

I'm not entirely sure what the correct response should be, but I think that
the EC might not be doing the right thing either. In any event, this keeps
it from getting stuck.

BUG=none
BRANCH=none
TEST=manual

Try

  ectool test 8 14
  ectool test 0 14

The first command fails because we told it to, but the second command also
fails because the EC is expecting something else.

After this change, the second command works.

Change-Id: I513294396cff872018316c354d2c41953eb6fdf6
Signed-off-by: Bill Richardson <wfrichar@chromium.org>