Update patch set 1
Patch Set 1:

This only happens when TSU6721 fails to detect the charger and reports non-standard charger, so not every second. As for why we need this, when the system is off, TSU6721 sometimes misdetect a video dongle as non-standard charger after it first recognize the dongle as USB host. This confuses the EC. What's worse, when the EC updates the device type, it reads non-standard charger from TSU6721 and thus drop video dongle from detected type. The bug is that when this happens, ID_MUX is left high and TSU6721 never sees the ID pin again.

Patch-set: 1