Update patch set 1
Patch Set 1:

Another thing that you could do would be to wait for some amount of time, like 3 seconds, and then just poll the cable_det once.  The analogix driver is also going to debounce this signal.  If it sees the signal high for long enough (currently 1 second), it will power up the chip.  Once that happens, cable detect is guaranteed to stay high as long as the driver still thinks the bridge is active.  After that, it will de-power the chip.

Patch-set: 1
Label: Commit-Queue=0