Update patch set 1
Patch Set 1: (4 inline comments)

A couple of nits on naming, and a couple of questions on the flash_ec part.

While not absolutely necessary IMO it might be easier to break this up into three CLs, one to add in the board/puppy/ support, one to modify the #ifdefs needed in the generic code, and one to add support to the flasher.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0