Update patch set 1
Patch Set 1: (1 inline comment)

Please let me know if there is a cleaner way to call the board-specific functions to enable/disable the discharge on AC mode. Perhaps I should try to support this on all boards but have it be a no-op if the board doesn't support it?

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0