Update patch set 2
Patch Set 2: Do not submit

Hopefully we'll be able to refactor the code so we don't fork the x86-haswell power sequencing code for each board.  There's a lot of code in there which is common across all boards, and maintaining it across multiple copy/paste/modify versions will be unpleasant.

(For working code during bringup, though, this is ok, since it frees you from needing to worry about breaking the other platforms.)

Marking this -2 so it doesn't accidentally get submitted in this form.

Patch-set: 2
Label: Commit-Queue=0