Update patch set 2
Patch Set 2:

It really needs to be forked. The reality is that the GPIO names are not the same and some don't even exist on certain boards. Moreover, some sequencing is done in hw on some boards while completely GPIOs on others.

It might be possible to try to refactor things, but that is much longer term. The GPIO strings should reflect the signal names on the board, and there shouldn't be a lot of #ifdef to handle the case of boards not even implementing certain things. My current take is that the reality of current boards is that they are not all the same.

Patch-set: 2
Reviewer: Gerrit User 1002124 <1002124@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0