Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(4 inline comments)

I'll buy that the GPIOs are different.

But there's a lot of functionality which is the same across all the x86 platforms, which should be refactored out into power_x86_common.c or something like that.  I hadn't done that for link because there was only one platform, and Bill hadn't done it for Slippy/Falco because he was rushed.

But if we're going to make another fork, it's time.  I can work on that in parallel with your bolt bringup.  

What I want to avoid is Peppy and other to-be-named variants doing the same thing, and then when I go to do the refactoring there are half a dozen files I need to change instead of 2, and nobody wants to accept the changes because at any point in time one of the projects is too close to shipping.

So, either we should check in this CL (so I can keep bolt building with my refactoring) or I can work on the ivybridge/haswell files and you can port the changes across to this CL yourself.  Up to you.

(also, side note: please cc Bill on all the Bolt CLs, since Falco may need similar changes)

Patch-set: 2
Label: Code-Review=-1