haswell: Enable LTE module in S0
Note that the PP3300_LTE_EN signal does different things on
different boards. On peppy/slippy it controls both LTE power
and gates a wake interrupt to the PCH. On falco it just gates
the wake interrupt; module power is tied to the DX rail.

On all boards there is a separate DISABLE_L signal from the PCH.

BUG=chrome-os-partner:20513
BRANCH=falco,peppy
TEST=Manual. Verify module detectable via lsusb in S0. Verify power
to module is disabled in S3.

Signed-off-by: Dave Parker <dparker@chromium.org>

Change-Id: I4984081009e4a1ce8ad8996e97f779c545829ce5