Update patch set 1
Patch Set 1: Looks good to me, approved

It's probably worth noting that this blindly enables a regular regardless of knowing the presence of a modem or not. This can lead to more power used even if there is no modem.  We need to address this later when we can determine the SKU of the final board (probably need an EC command to do this for us).

Patch-set: 1
Label: Code-Review=+2
Label: Commit-Queue=0