Update patch set 1
Patch Set 1: (5 inline comments)

Thanks for the comments. Please see patchset 3. (patchset 3 just fixed a copyright date in a header file. The main fixes are in patchset 2).

This appears stable on both falco and peppy. I think my earlier problems with the watchdog timing out were due to either my i2c probewires shorting the data and clock together and/or not removing the build directory for a board when switching branches before building.

Patch-set: 1