Update patch set 3
Patch Set 3: (2 inline comments)

We have never duplicated the the current chipset_haswell for all the haswell boards. They are all so similar that it wasn't necessary. As for GPIO names, I understand that isn't a requirement. However, if you ware working on a board I consider it a plus that the SW names match with the hardware names. No extra translation step.

chipset_haswell.c should probably be named chipset_haswell_ref.c to indicate that it's based off of a reference design. It is not the defacto haswell way as many things can be changed from board to board.

You can drop the suswarn->susack if you want in the haswell code as those boards don't actually have the resistors to make that happen stuffed.

Patch-set: 3