Falco: AC adapter identification command
This adds a function to identify the AC adapter. We'll need this to set
various battery charging limits, based on the capability. This only applies
to Falco, AFAIK.

BUG=chrome-os-partner:19594
BRANCH=none
TEST=manual

Plug in various approved adapters. From the EC console, run "adapter". It
should tell you the rated power for each one.

Change-Id: Id6d142fa81f20ec9233b0faa2fcb1d53cf7b7ef5
Signed-off-by: Bill Richardson <wfrichar@chromium.org>