Turn off boost when the battery is nearly full
Boost is consuming about 1W. To reduce power draw in S5, we should turn
off boost when the battery is nearly full. The battery discharges slowly
in this case, so we also need to turn boost back on when the battery
level drops too low.

BUG=chrome-os-partner:20768
TEST=In S5, see boost disabled when battery is charged to 98%. And see
boost enabled when battery drops to 94%.
TEST=When boost is disabled, power on the system. Check boost is
enabled.
BRANCH=Spring

Change-Id: Ibc6a8af17bb6ce419abbcc8b693b7f90bf59839c
Signed-off-by: Vic Yang <victoryang@chromium.org>