Make fancy trickle charging logic optional
We implemented a fancy state machine for link to try and feed
almost-dead batteries smaller amounts of current than the charger
could normally supply, by dithering down the voltage to less than was
requested.  It's unclear now that this was the right thing to do; it
would be a lot simpler just to give the battery the smallest non-zero
amount of current we can give it at the voltage it asks for.

This changes removes the precharge logic from all newer platforms
while retaining it (for now) for link.

Alternately, we could just delete the code, since link has already
branched.

BUG=chrome-os-partner:20881
BRANCH=none
TEST=find a really dead batery and try to charge it, or do what I did
     and just hack the battery module to lie about the requested values,
     and see that it always gets the minimum current step at the requested
     voltage.

Change-Id: I30bcb62bbe215abafb3a98c52e2afed3b74ead1a
Signed-off-by: Randall Spangler <rspangler@chromium.org>