Update patch set 2
Patch Set 2: Abandoned

The original motivation for this change has been fixed with a hardware change. Dropping this for now.

However, I think we should consider that for some Chromebooks, the brief AP boot with the lid closed is noticeable by the user due to the power LED turning on...and currently the LCD initializing...but that might be due to developer mode.

Patch-set: 2