Update patch set 3
Patch Set 3:

What's the status of this CL?  I'd posted it originally to give Aaron a starting point for bringup.  Looks like he's improved it since then.  I'm guessing that the cleanup I've done to the x86 platforms will require some cleanup/rebasing.

Are you doing that in this CL, or do you want me to abandon this CL and you'll create a fresh one when you're ready?

Thanks

Patch-set: 3
Label: Commit-Queue=0