Update patch set 1
Patch Set 1: Ready

Yeah, it's a little damned-if-you-do-damned-if-you-don't.

The problem I'm trying to solve is that CONFIG_ macros can get pulled in from a couple places, and if you #include the wrong file, you silently don't get any of the macros.  That's led to code which did

#include "board.h"

#ifdef CONFIG_FOO  // actually defined in chip/$(CHIP)/config.h

which silently disables a feature.

Patch-set: 1
Label: Commit-Queue=+1