Update patch set 4
Patch Set 4: Ready

Tested on Falco, except for lid switch because the proto board with servo doesn't have a lid switch.

Lid switch code is identical between Link and Falco and it works on Link.  It should also be pretty darn obvious if the system stops booting on lid open on Falco...

Patch-set: 4