Fix extracting CONFIG tokens from config.h
We need to use cpp -dM instead of -dN, because -dN doesn't process #undef
statements.  That is, if you have
"  #define CONFIG_FOO"
"  #undef CONFIG_FOO"
then -dN will happily print both lines, which results in CONFIG_FOO being
defined when processing the makefile.  -dM processes the defines and undefs
so that the resulting list of macros only includes those that are defined.

We didn't notice this before, because we temporarily commented out
config statements instead of #undef'ing them - or the configs which
were undef'd only resulted in conditional non-compilation of pieces of
code inside a file instead of the whole file not being compiled.

This change also tidily alphabetizes the resulting configs.

It also better filters CONFIGs to ensure that
"  #define __CROS_EC_CONFIG_CHIP_H"
does not show up as a CONFIG_CHIP_H token, as it did previously.

BUG=chrome-os-partner:20985
BRANCH=none (though may be needed if any future cherry-picked change requires
             undefining a CONFIG)
TEST=Add #undef CONFIG_ADC to the end of config.h and see that the ADC module
     is not compiled.  Also helps to add
         $(info $(_flag_cfg))
     to the Makefile to see the list of tokens, and verify that CONFIG_ADC is
     not among them.

Change-Id: I18db60099e87857473ba54c3c9fff1116d4f1a93
Signed-off-by: Randall Spangler <rspangler@chromium.org>