Update patch set 3
Patch Set 3:

Bleh, what I meant to say is that we should have one place where host software decides how much to send. Right now we have 2, the legacy write_block_size and the protocol max parameter size, neither of which are optimal. We should simplify this down to 1 and make it give us the info we actually want to use.

Patch-set: 3