Update patch set 4
Patch Set 4:

Sounds good.  How exactly does servo get this in an unmodified board?  Wouldn't it need direct access to the i2c line that the battery is on?  I don't see it on the schematics.

The goal here is to check the battery without running any code on the AP (so we don't affect the results).  SLSI has said that polling too often on the AP was actually increasing the system load and causing changes in the power usage, especially at low CPU frequencies.

If this functionality is somewhere else I'm happy to abandon, though!  :)

Patch-set: 4
Label: Commit-Queue=0
Label: Verified=0