Add charger_falco.c and tests for it.
The falco charger can run in a "turbo" mode, which lets it draw from the
battery to provide extra power to the AP in short bursts. In order for this
to work properly, the EC has to watch the current closely to make sure
specific limits are observed. It also has to recognize specific adapters,
since those limits vary depending on the rated power that the adapter can
provide.

This adds the basic functionality, plus a test for it.

BUG=chrome-os-partner:20739
BRANCH=falco,peppy
TEST=manual

  make BOARD=${BOARD} runtests

On Falco, you can also use the "adapter" EC command to see what's going on.
Try replacing the adapters and running that command to see how they're
identified, too:

  > adapter
  Adapter 65W (590mv), turbo 1, AP_throttled 0
  >

Change-Id: I1a974f59ccb49ffd9170ea4500a73fbf6f10dfbc
Signed-off-by: Bill Richardson <wfrichar@chromium.org>