spring: Cut off the battery when its charge is critically low
When the battery charge is too low, we should cut off the battery to
prevent it from being depleted. This allows us to ignore the case where
the battery voltage is 0V.

Also, if we are already in hibernation, wake up every hour to check
battery cut off state.

BUG=chrome-os-partner:21107
TEST=Unplug the AC. Wait for the battery to cut off.
BRANCH=Spring

Change-Id: Ie69a89b4a01db0c5435681f0431d31a99b5014c3
Signed-off-by: Vic Yang <victoryang@chromium.org>