Update patch set 1
Patch Set 1:

(This highlights a weakness in the EC tests.  We need some test commands that send/receive maximum-length packets, and we need a better way to verify that a range of host commands still work properly after a change - not just that a single malfunctioning command works properly.  Without that, it's really tricky to make changes to the protocol stack without introducing new bugs.)

Patch-set: 1