spring: support pull-down and pull-up on ENTERING_RW
Depending on the version of the board, the Silego GreenPakII has an
internal pull-down (older boards) or an internal pull-up on the pin
connected to the ENTERING_RW net, and the active edge is inverted (from
raising edge to falling edge).
Ensure we support both versions by probing the default level on the
ENTERING_RW GPIO and doing a pulse in the right direction.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:21396
TEST=On both hardware versions of Spring, use either the RO EC or the RW EC,
and ensure we are reading back the right level on the EC_IN_RW GPIO on
the AP side (GPD1_7).

Change-Id: I5bf45edd5bf7d0ed65f3ae25a8abe204bbed3989