spring: record TPS65090 reset
When we are doing a TPS65090 hard reset to recover from a bad situation,
we record a bit in the backup memory.
This information will be available in the chip revision information,
it will return "TPSRESET" if the bit has been flipped.

The bit is never cleared, but the "scratchpad" command is now activated
on the EC console : doing "scratchpad 0" on the EC serial port will
reset it.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:21107
TEST=on Spring, run "ectool chipinfo | grep revision"
before and after having run "pmu reset" on the EC console.

Change-Id: I9ebdeba59b978cd974b0ec268cba698b8cab0bf9