spring: Do not allow power input when cutting off the battery
If a charger is plugged in while we are cutting off the battery, the EC
will be powered by external power, and the battery will be in a state
where it doesn't charge. Let's disable power input in this case, so that
the EC will die after the battery cut itself off. The input power then
resume to revive both the EC and the battery.

BUG=chrome-os-partner:21107
TEST=Plug in the charger while cutting off the battery.
BRANCH=Spring

Change-Id: I5505485be7872f9887efc725a2b3aa610c6b42b9
Signed-off-by: Vic Yang <victoryang@chromium.org>