stm32: allow to go into stop mode during suspend
The SLEEP_MASK_AP_RUN bit in the sleep mask is mainly used to avoid to
go to stop mode while the AP can send us an I2C request (as the STM32
I2C controller does not wake the SoC from stop mode).
When the AP is in suspend, it cannot send us a request, we can remove
that mask bit. As a consequence, the EC might go into stop mode while
the AP is in S3 (if we are not charging...).

Also remove the sleep locking done in the charger state machine, the
resources incompatible with STOP mode (e.g. ilim PWM) should be
protected by finer grained sleep mask bits.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=none
TEST=on Spring, ensure we suspend/resume properly and measure S3 power.

Change-Id: I029feec6f6e1978245763d89a666a876f6b7cf58