Reduce board-specific init functions
board_config_pre_init() is only needed to configure GPIOs on STM32F,
so rename it to gpio_board_pre_init(), call it from inside the STM32F
gpio_pre_init() implementation, and only allow it for CHIP_FAMILY_stm32f.

board_config_post_gpio_init() is no longer used and can be removed.

EC_INT was set up in both the pre-init and in the GPIO list; remove
the redundant configuration.

This is a refinement along the way to implementing
gpio_set_alternate() on STM32F.

BUG=chrome-os-partner:21618
BRANCH=none
TEST=build all boards; boot spring

Change-Id: I320e46dda3bc56605468a0df88730da6d7500450
Signed-off-by: Randall Spangler <rspangler@chromium.org>