Fix leaving keyboard scanning disabled on brief power button press
If the power button is pressed for a shorter period than the debounce
timeout, then the debounced state never changes.  This was causing the
power button state machine to disable scanning (in the interrupt
handler) but never re-enable it (in the deferred handler).

Easy fix; just re-enable based on whether the current state is
released, not whether the debounced state is transitioning to
released.

BUG=chrome-os-partner:21772
BRANCH=all (falco, pit, etc.)
TEST=type on console.  briefly flick power button.  type more; should work.

Change-Id: I9723a6aa10f122fcee62702b85ce7981b1c8103a
Signed-off-by: Randall Spangler <rspangler@chromium.org>