Haswell: shutdown directly to G3  without pausing in S5
We've been pausing in S5 for ten seconds for various arcane reasons related
to clock rates and USB peripherals. We don't need to do that anymore, and
there are other arcane reasons why it's better if we don't.

BUG=chrome-os-partner:21791
BRANCH=falco,peppy
TEST=manual

On the EC console, limit the output to just the chipset channel:

> chan 4

Now boot the AP, then shut down.

Before you'd see a ten-second pause in S5, like this:

  [29.586858 x86 power state 3 = S0, in 0x00df]
  [29.587268 x86 power state 7 = S0->S3, in 0x009f]
  [29.587707 x86 power state 2 = S3, in 0x009f]
  [29.587959 x86 power state 8 = S3->S5, in 0x009f]
  [29.588474 x86 power state 1 = S5, in 0x009c]
  [29.588733 x86 power state 1 = S5, in 0x009c]
  [29.603317 x86 power state 1 = S5, in 0x0094]
  [39.603612 x86 power state 9 = S5->G3, in 0x0094]
  [39.604137 x86 power state 0 = G3, in 0x0000]
  [39.604376 x86 power state 0 = G3, in 0x0000]

With this change the pause is gone:

  [26.764160 x86 power state 3 = S0, in 0x00df]
  [26.764570 x86 power state 7 = S0->S3, in 0x009f]
  [26.765011 x86 power state 2 = S3, in 0x009f]
  [26.765262 x86 power state 8 = S3->S5, in 0x009f]
  [26.765777 x86 power state 9 = S5->G3, in 0x009c]
  [26.766220 x86 power state 0 = G3, in 0x0008]
  [26.766526 x86 power state 0 = G3, in 0x0008]
  [26.770517 x86 power state 0 = G3, in 0x0000]

Change-Id: I05e19ddfe9dfa1bcc2a29103d120910c4371b88e
Signed-off-by: Bill Richardson <wfrichar@chromium.org>