stm: i2c: add unwedge function and command
BUG=chrome-os-partner:19286
TEST=Manual test on peach pit. First, wedge the i2c bus. A simple pull-down on
SCL does it. Or keep executing the "battery" command on the EC console while
busy running flashrom on the host to read back the EC until the "battery"
command fails. Next, run the "i2cunwedge" command from the EC console. After
that, run the "battery" command again and observe that the error goes away.

Note that most of time, running "i2cunwedge" once fixes the problem. But
sometimes, "i2cunwedge" may be needed to run again for the bus to recover.
BRANCH=pit

Change-Id: I78846f9f72d800426c5926aae04d046a48c1a3ea
Signed-off-by: Hung-ying Tyan <tyanh@chromium.org>