stm: i2c: try unwedging the bus when i2c_xfer fails
BUG=chrome-os-partner:19286
TEST=Manual test on peach pit. Wedge the bus with a simple pull-down on SCL,
or keep executing the "battery" command on the EC console while busy running
flashrom on the host to read back the EC until the "battery" command fails.
Observe that i2c_xfer reports error and then unwedging kicks in automatically.
After that, execute the "battery" command again and see it succeed.
BRANCH=pit

Change-Id: I628ec9ab79a3169ed8afc4e3ac69e3fa6eed3139
Signed-off-by: Hung-ying Tyan <tyanh@chromium.org>