i2c: add wedge command
This command wedges the I2C bus by writing part of a byte to or reading part
of the response from the slave device.

To enabled the wedge command you must define CONFIG_CMD_I2CWEDGE and you must
pick gpio pins and an i2c port to use the wedge command on such as:

BUG=chrome-os-partner:19286
TEST=Manual test on peach pit. Define config in board/pit/board.h to enable
the command:

On the EC console, execute the following "i2cwedge" command

   i2cwedge 0x90 0 1    (wedge write)
or
   i2cwedge 0x90 0 2    (wedge read)

and then "battery". Observe that the command reports an error.

Similarly, execute

   i2cwedge 0x90 0 5    (wedge write + reboot)
or
   i2cwedge 0x90 0 6    (wedge read + reboot)

on the EC console and observe a reboot. Then execute "battery" and observe
that the command works properly.
BRANCH=none

Change-Id: I10ccb21b047df907a4dfdbd84c0f582cfa2d939a
Signed-off-by: Hung-ying Tyan <tyanh@chromium.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>