pit: i2c: add wedge command
This command wedges the I2C bus by reading only part of the response
from the PMU.

BUG=chrome-os-partner:19286
TEST=Manual test on peach pit. Execute the following "i2cwedge" command

   i2cwedge 0x90 0 1    (wedge write)
or
   i2cwedge 0x90 0 2    (wedge read)

and then "battery" on the EC console. Observe that the command reports an
error.

Execute

   i2cwedge 0x90 0 5    (wedge write + reboot)
or
   i2cwedge 0x90 0 6    (wedge read + reboot)

on the EC console and observe a reboot. Then execute "battery" and observe
that the command reports an error.
BRANCH=pit

Change-Id: I10ccb21b047df907a4dfdbd84c0f582cfa2d939a
Signed-off-by: Hung-ying Tyan <tyanh@chromium.org>
Signed-off-by: Doug Anderson <dianders@chromium.org>