Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

While this may not hurt, it doesn't seem like it's solving the actual problem at hand.  

I _think_ that your wedge command wedges the bus because it reconfigures the pin to no longer be an i2c pin--it's now a GPIO.  It's not surprising that we can't talk on the bus after you do that.  ...but it's very unlikely that was what was really happening to the bus in cases that we were wedged.

This fix undoes the badness that your wedge command did, but I don't think it's likely to fix other wedges.

Patch-set: 2
Label: Commit-Queue=0
Label: Verified=0