pit: Only enable FET4 by default at cold reset
Now that U-Boot and kernel can properly talk to the EC in pit, there's
no reason to hack all the FETs on.  We only need to turn on FET4 which
enables SD card booting.

We'll leave the old "all fets on" hack there for "puppy", though.
Apparently that still needs it?

BRANCH=pit
BUG=chrome-os-partner:21975
TEST=Boot up and see LCD turn on.
TEST=Cold reset while holding recovery and can boot from SD card.

Change-Id: Iae96375ac7bd1a9eed8243367332cf003b62c48d
Signed-off-by: Doug Anderson <dianders@chromium.org>