Update patch set 2
Patch Set 2: No score

I didn't realize that this was actually tested against the constant execution of the battery command and that it actually fixed it.  I think we could take something like this, though perhaps we should still do a more full unwedge once we can reproduce a more full wedge.

Patch-set: 2