Update patch set 1
Patch Set 1:

Also: one option could be to keep this as a patch in gerrit and don't actually land it.  It's not common to need to wedge the bus, so maintaining it in the source code (and paying the cost of keeping it in the ROM) seems a little silly?

Patch-set: 1