Peppy: Enable PP5000 while checking initial keyboard state
BRANCH ONLY! Not for ToT.

Active ESD components on the keyboard scan matrix require 5V
power. These components must be powered to read the keyboard
state correctly. Poll the keyboard for up to 5mSec after
powering the 5V rail to wait for the keyboard state to settle
down.

BUG=chrome-os-partner:22127
BRANCH=peppy
TEST=Manual.
  Boot normally.
  Boot normally with Esc + F3 + Power + another key.
  Boot into recovery with Esc + F3 + Power.

Change-Id: I064cf6eb98b47c897079f145e945b4ab3ae4ef23
Signed-off-by: Dave Parker <dparker@chromium.org>