8042: Wake the keyboard scan task when keystrokes are enabled
Since the keyboard_scan_task is the only place that calls
keyboard_raw_enable_interrupt(1) it is possible for the keyboard
to be left with interrupts disabled/pending if the keyboard scan
task is asleep.

This change will wake the keyboard_scan_task when the 8042
keystroke_enable(1) is called, which will wake up the task and
it will call into enable the keyboard interrupt and clear pending
interrupts.

I thought keyboard_enable() might be a better place to do this
but that seems to cause a hung task when flashing firmware...

This condition happens semi-regularly on resume on Falco and Peppy
devices. There may be a better approach here so for now I am only
commiting this to the Falco/Peppy firmware branch but a similar
or better fix is probably needed in TOT.

BUG=chrome-os-partner:22169
BRANCH=falco,peppy
TEST=manual: many suspend/resume cycles on falco with the physical
lid to ensure that the keyboard is still functional.

Change-Id: I2f56f9c07b852c4b03b090d0da6b7128804962ab
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>