Update patch set 4
Patch Set 4:

Sorry, I needed to rework this and add the polling loop back. 

Logs from the debug version showed that at least one trip through the loop was needed (~1msec) for the keyboard to settle down after the 5V rail was powered. I've set the polling timeout to 5mSec to be on the safe side. I moved the hack into a new peppy-specific function that returns the initial stable state of the keyboard.

Patch-set: 4
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0