Update patch set 1
Patch Set 1: Code-Review+2

This is a step in the right direction.

Unclear how much stuff we want to live in the board/ directories, since we tend to copy-paste-modify to create new boards.  In this case it might be ok.

Patch-set: 1
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2