Update patch set 1
Patch Set 1: Commit-Queue+1 Verified+1

I'm merging this so that this is not confusing for Kirby.

As for board specific stuff concern, we do have more and more files in common/ with board names. For those that might be able to be reused (like extpower_falco), that's probably the right place for them. But for those with very board-specific information (e.g. extpower_spring), maybe we need to move them so that we don't have so many of them when we have more projects?

Patch-set: 1
Label: Commit-Queue=+1
Label: Verified=+1