Update patch set 4
Patch Set 4:

Ok to have this for now, but we probably eventually want this under AP control in the bridge driver (so the switcher can be shut down if the back light is off to save power; we will need to measure how much that is). The EC must force the backlight off when the lid is closed (or after 12 sec to match spring) as a safety cutout against overheating if the AP does not correctly shutdown.

Patch-set: 4
Reviewer: Gerrit User 1001164 <1001164@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0